redux file structure refactor
