+++ src/index.tsx
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux'
 import { BrowserRouter as Router } from 'react-router-dom'
 import './index.css'
 import App from './App'
-import store from './redux/store'
+import store from './store'
 import * as serviceWorker from './serviceWorker'
 
 ReactDOM.render(
